Added arc.present package which will contain all formatters and ui components moving foward 	- Got rid of arc.ui for now, going to pull in Scythe's ui stuff at some point 	- Moved formatters into arc.present 	- Added the option to change the color of the Box presenter border
