--- src/arc/ui/keys.py
@@ -1,79 +1,0 @@
-A = 65
-B = 66
-C = 67
-D = 68
-E = 69
-F = 70
-G = 71
-H = 72
-I = 73
-J = 74
-K = 75
-L = 76
-M = 77
-N = 78
-O = 79
-P = 80
-Q = 81
-R = 82
-S = 83
-T = 84
-U = 85
-V = 86
-W = 87
-X = 88
-Y = 89
-Z = 90
-
-a = 97
-b = 98
-c = 99
-d = 100
-e = 101
-f = 102
-g = 103
-h = 104
-i = 105
-j = 106
-k = 107
-l = 108
-m = 109
-n = 110
-o = 111
-p = 112
-q = 113
-r = 114
-s = 115
-t = 116
-u = 117
-v = 118
-w = 119
-x = 120
-y = 121
-z = 122
-
-ZERO = 48
-ONE = 49
-TWO = 50
-THREE = 51
-FOUR = 52
-FIVE = 53
-SIX = 54
-SEVEN = 55
-EIGHT = 56
-NINE = 57
-
-NUMBERS = (ZERO, ONE, TWO, THREE, FOUR, FIVE, SIX, SEVEN, EIGHT, NINE)
-
-OPENING_BRACKET = 91
-BACKSLASH = 92
-CLOSING_BRACKET = 93
-
-ESC = 27  # \u001b
-UP = (ESC, OPENING_BRACKET, A)
-DOWN = (ESC, OPENING_BRACKET, B)
-RIGHT = (ESC, OPENING_BRACKET, C)
-LEFT = (ESC, OPENING_BRACKET, D)
-
-# Enter is 13 on some systems and 10 on others
-ENTER_TUP = (13, 10)
